+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 
-## 0.10.0
+## 1.1.0
+
+ * Added new AWS config option
+ * Updated cleaning to be less aggressive to allow overlapping paths in config
+ * Added separate `clean` 
+
+## 1.0.0
 
  * Added Mustache templating
  * Removed `html-concat` option